+++ tests/testthat/test-dbpedia.R
@@ -3,14 +3,15 @@ test_that(
   {
 
     withr::local_package("quanteda")
-    quanteda_corpus_min <- corpus(quanteda::data_char_ukimmig2010[1:3])
+    quanteda_corpus_min <- quanteda::corpus(quanteda::data_char_ukimmig2010[1:3])
 
     uritab <- get_dbpedia_uris(
       x = quanteda_corpus_min,
       api = "http://api.dbpedia-spotlight.org/en/annotate",
       language = "en",
       progress = FALSE,
-      verbose = FALSE
+      verbose = FALSE,
+      retry = 0L
     )
   }
 )